+++ scripts/ci-opencv.sh
@@ -54,6 +54,7 @@ CORE_TEST_PATTERNS=(
   '*Core_Transpose*'
   '*Core_MinMaxLoc*'
   '*Core_ConvertScale*'
+  '*Core_Exp*'
 )
 CORE_TEST_PATTERNS_STR="$(join_strings_with_colon "${CORE_TEST_PATTERNS[*]}")"
 ../../../conformity/opencv_kleidicv/bin/opencv_test_core \